Grabbing keyboard pulse data using Whatpulse API
This PHP script will use the Whatpulse API to loop through all
historical pulses and display your pulse data on a page.

Data can be saved to a mySQL database or CSV file for further analysis,
I just wrote this script because there was no default one to be found
online to save all historical pulses.

I have made the assumption that you will not pulse more than 100 times
in 2 months, this is because the API limits 100 queries at a time - feel
free to change this if you pulse a lot.

Also the API doesn't seem to retrieve pulse data prior to August 2015...
nothing I an do about that i'm afraid, it is a shame since I have been a
member since January 2008.